[style]메인,검색결과페이지 스타일컴포넌트 분리
