+++ src/pages/searchResults/SearchResults.tsx
@@ -1,17 +1,12 @@
 import React, { useCallback, useEffect, useState } from 'react';
-import { Link, Navigate, useLocation, useNavigate } from 'react-router-dom';
+import { useLocation, useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import styled, { css } from 'styled-components';
+import * as St from '../../styles/searchresults/SearchResultsStyle';
 import { setSearchResults } from '../../redux/modules/searchSlice';
-import { FaArrowRight } from 'react-icons/fa';
 import { FaArrowDown } from 'react-icons/fa6';
 import { FiArrowUp } from 'react-icons/fi';
-import { FaCheckSquare } from 'react-icons/fa';
 import { researchItems, ResearchResults } from './researchItem';
 import Dropdown from '../../styles/searchresults/Dropdown';
-import { divide, sortBy } from 'lodash';
-import { Communityy, UsedItem } from '../home/usedtypes';
-import * as St from '../../styles/products/productsList/StProductsCard';
 import CommunityList from '../../components/community/CommunityList';
 import { RootState } from '../../redux/store/store';
 import ProductsCard from '../../components/prducts/ProductsCard';
@@ -181,52 +176,52 @@ const SearchResults: React.FC = () => {
 
   return (
     <>
-      <SearchResultsContainer>
-        <SearchResultsCountContainer>
-          <Icon />
-          <FullCounts>
+      <St.SearchResultsContainer>
+        <St.SearchResultsCountContainer>
+          <St.Icon />
+          <St.FullCounts>
             {isLoading
               ? '검색 중...'
               : usedItemCount === 0 && communityCount === 0
               ? '해당 검색어에 대한 결과를 찾을 수 없어요'
               : `${usedItemCount + communityCount}개의 결과가 검색되었어요`}
-          </FullCounts>
-        </SearchResultsCountContainer>
-        <ResultListContainer>
-          <UsedItemResultsContainer>
+          </St.FullCounts>
+        </St.SearchResultsCountContainer>
+        <St.ResultListContainer>
+          <St.UsedItemResultsContainer>
             {/* TAB */}
             {isMobile ? (
-              <CountBar>
-                <CountPost>
-                  <ProductsCount
+              <St.CountBar>
+                <St.CountPost>
+                  <St.ProductsCount
                     showClickedList={showClickedList}
                     onClick={() => handleTabClick('중고물품')}
                   >
                     중고거래({usedItemCount})
-                  </ProductsCount>
-                  <CommunityCount
+                  </St.ProductsCount>
+                  <St.CommunityCount
                     showClickedList={showClickedList}
                     onClick={() => {
                       handleTabClick('커뮤니티');
                     }}
                   >
                     커뮤니티({communityCount})
-                  </CommunityCount>
-                </CountPost>
+                  </St.CommunityCount>
+                </St.CountPost>
                 <Dropdown
                   isOpen={isOpen}
                   setIsOpen={setIsOpen}
                   clickMenu={clickMenu}
                   setClickMenu={setClickMenu}
                 />
-              </CountBar>
+              </St.CountBar>
             ) : (
               // 데스크탑
-              <CountBar>
-                <CountList>
+              <St.CountBar>
+                <St.CountList>
                   {usedItemCount}개의 상품이 거래되고 있어요
-                </CountList>
-                <LinktoUsedProducts
+                </St.CountList>
+                <St.LinktoUsedProducts
                   onClick={handleToggleShowAllProducts}
                   style={{
                     display:
@@ -238,8 +233,8 @@ const SearchResults: React.FC = () => {
                   {showAllProducts ? <p>숨기기</p> : <p>전체보기</p>}
 
                   {showAllProducts ? <FiArrowUp /> : <FaArrowDown />}
-                </LinktoUsedProducts>
-              </CountBar>
+                </St.LinktoUsedProducts>
+              </St.CountBar>
             )}
             {/* 검색 결과 */}
             {isMobile && !showClickedList && (
@@ -247,7 +242,7 @@ const SearchResults: React.FC = () => {
             )}{' '}
             {!isMobile && (
               // 중고 데스크탑
-              <ProductsProtecter>
+              <St.ProductsProtecter>
                 {/* productsCard 컴포넌트 props넘기기 */}
 
                 {isLoading ? (
@@ -255,16 +250,16 @@ const SearchResults: React.FC = () => {
                 ) : (
                   <ProductsCard posts={productsData} />
                 )}
-              </ProductsProtecter>
+              </St.ProductsProtecter>
             )}
-          </UsedItemResultsContainer>
+          </St.UsedItemResultsContainer>
 
           {/* ////////////////////////////////////////////////////////////////////////////////////////////////////////// */}
           {/* 커뮤니티 */}
-          <CommunityResultsContainer showClickedList={showClickedList}>
-            <CommunityTitle showClickedList={showClickedList}>
-              <CountList>{communityCount}개의 이야기가 있어요</CountList>
-              <LinktoCommunityPosts
+          <St.CommunityResultsContainer showClickedList={showClickedList}>
+            <St.CommunityTitle showClickedList={showClickedList}>
+              <St.CountList>{communityCount}개의 이야기가 있어요</St.CountList>
+              <St.LinktoCommunityPosts
                 onClick={handleToggleShowAllCommunity}
                 style={{
                   display: communityResults.length <= 6 ? 'none' : 'flex'
@@ -273,8 +268,8 @@ const SearchResults: React.FC = () => {
                 {showAllCommunity ? <p>숨기기</p> : <p>전체보기</p>}
 
                 {showAllCommunity ? <FiArrowUp /> : <FaArrowDown />}
-              </LinktoCommunityPosts>
-            </CommunityTitle>
+              </St.LinktoCommunityPosts>
+            </St.CommunityTitle>
             {/* 커뮤니티 모바일 */}
             {isMobile && showClickedList && (
               <CommunityList posts={sortedCommunityResults} />
@@ -286,604 +281,11 @@ const SearchResults: React.FC = () => {
               ) : (
                 <CommunityList posts={slicedCommunityResults} />
               ))}
-          </CommunityResultsContainer>
-        </ResultListContainer>
-      </SearchResultsContainer>
+          </St.CommunityResultsContainer>
+        </St.ResultListContainer>
+      </St.SearchResultsContainer>
     </>
   );
 };
 
 export default SearchResults;
-
-const SearchResultsContainer = styled.div`
-  display: flex;
-  max-width: 144rem;
-  flex-direction: column;
-  min-height: 100vh;
-  margin: 0 auto;
-`;
-
-const SearchResultsCountContainer = styled.div`
-  margin: 0 auto;
-  margin-top: 6rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  @media screen and (max-width: 768px) {
-    margin-top: 2rem;
-  }
-`;
-const Icon = styled(FaCheckSquare)`
-  margin: 0 auto;
-  width: 6.6rem;
-  height: 6.6rem;
-  color: var(--opc-100);
-  @media screen and (max-width: 768px) {
-    width: 4.4rem;
-    height: 4.4rem;
-  }
-`;
-const FullCounts = styled.div`
-  margin-top: 2rem;
-  font-size: var(--fontSize-H1);
-  font-weight: var(--fontWeight-bold);
-  @media screen and (max-width: 768px) {
-    font-size: var(--fontSize-H4);
-  }
-`;
-const ResultListContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  margin: 0 auto;
-  margin-top: 8rem;
-  @media screen and (max-width: 768px) {
-    width: 100%;
-    margin-top: 1rem;
-  }
-`;
-
-const ProductsProtecter = styled.div`
-  margin-top: 2rem;
-  @media screen and (max-width: 768px) {
-    margin-top: 1rem;
-  }
-`;
-
-const UsedItemResultsContainer = styled.div`
-  margin-top: 2rem;
-  width: 77.5%;
-  margin: 0 auto;
-  margin-bottom: 4rem;
-  @media screen and (max-width: 768px) {
-    width: 93%;
-    min-width: 32rem;
-    margin-top: 2rem;
-  }
-`;
-const CountBar = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  margin: 0 auto;
-  margin-top: 2rem;
-  @media screen and (max-width: 768px) {
-    margin-top: 1rem;
-  }
-`;
-const CountPost = styled.div`
-  display: flex;
-  width: 100%;
-  font-size: var(--fontSize-H3);
-  align-items: center;
-  gap: 1rem;
-
-  @media screen and (max-width: 768px) {
-    width: 100%;
-    gap: 1rem;
-    font-size: var(--fontSize-H5);
-  }
-`;
-
-const ProductsCount = styled.h1<{
-  showClickedList: boolean;
-}>`
-  width: 40%;
-  cursor: pointer;
-  font-weight: ${({ showClickedList }) =>
-    !showClickedList ? 'var(--fontWeight-bold)' : 'var(--fontWeight-medium)'};
-  color: ${({ showClickedList }) =>
-    !showClickedList ? 'var(--opc-100)' : 'var(--black)'};
-`;
-const CommunityCount = styled.h1<{
-  showClickedList: boolean;
-}>`
-  width: 40%;
-  cursor: pointer;
-  font-weight: ${({ showClickedList }) =>
-    showClickedList ? 'var(--fontWeight-bold)' : 'var(--fontWeight-medium)'};
-  color: ${({ showClickedList }) =>
-    showClickedList ? 'var(--opc-100)' : 'var(--black)'};
-`;
-
-const CountList = styled.h1`
-  display: flex;
-  font-size: var(--fontSize-H3);
-  align-items: center;
-  @media screen and (max-width: 768px) {
-    font-size: var(--fontSize-H5);
-  }
-`;
-const LinktoUsedProducts = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 8.3rem;
-  height: 3.2rem;
-  text-decoration: none;
-
-  cursor: pointer;
-  font-weight: var(--fontWeight-bold);
-  gap: 0.8rem;
-  background: var(--opc-20);
-  border-radius: 4.5rem;
-  font-size: var(--fontSize-H6);
-  @media screen and (max-width: 768px) {
-    background: none;
-    width: 6rem;
-    font-size: 1.1rem;
-    gap: 0.3rem;
-  }
-  &:hover {
-    background-color: var(--opc-100);
-    color: var(--bgColor);
-  }
-  svg {
-    width: 1rem;
-    height: 0.9rem;
-    @media screen and (max-width: 768px) {
-      width: 1rem;
-      height: 0.9rem;
-      color: var(--opc-100);
-    }
-  }
-`;
-const UsedItemsList = styled.ul<{
-  usedItemCount: number;
-  showClickedList: boolean;
-}>`
-  width: 100%;
-
-  max-height: 100vh;
-  display: grid;
-  margin: auto;
-  grid-template-columns: repeat(5, 1fr);
-  margin-top: 2rem;
-  row-gap: 1.5rem;
-  column-gap: 1.8rem;
-  align-items: flex-start;
-  justify-content: center;
-  place-items: center;
-  margin-top: 2rem;
-
-  @media screen and (max-width: 1160px) {
-    grid-template-columns: repeat(4, 1fr);
-    margin-top: ${({ usedItemCount }) =>
-      usedItemCount !== 0 ? '5rem' : '2rem'};
-  }
-
-  @media screen and (max-width: 1024px) {
-    grid-template-columns: repeat(3, 1fr);
-    margin-top: ${({ usedItemCount }) => (usedItemCount !== 0 ? '5rem' : '')};
-  }
-
-  @media screen and (max-width: 768px) {
-    column-gap: 1.5rem;
-    row-gap: 1.8rem;
-    grid-template-columns: repeat(3, 1fr);
-    margin-top: ${({ usedItemCount }) => (usedItemCount !== 0 ? '3rem' : '')};
-  }
-  @media screen and (max-width: 670px) {
-    column-gap: 1.5rem;
-    row-gap: 1.8rem;
-    grid-template-columns: repeat(2, 1fr);
-    margin-top: ${({ usedItemCount }) => (usedItemCount !== 0 ? '2rem' : '')};
-  }
-  @media screen and (max-width: 520px) {
-    row-gap: 1.8rem;
-    grid-template-columns: repeat(2, 1fr);
-    margin-top: ${({ usedItemCount }) => (usedItemCount !== 0 ? '2rem' : '')};
-  }
-
-  @media screen and (max-width: 349px) {
-    grid-template-columns: repeat(1, 1fr);
-    margin-top: ${({ usedItemCount }) => (usedItemCount !== 0 ? '2rem' : '')};
-  }
-`;
-const ToProductsPage = styled.div`
-  text-decoration: none;
-  cursor: pointer;
-`;
-const ProductList = styled.li`
-  width: 20.8rem;
-  /* height: 31.5rem; */
-  display: flex;
-  flex-direction: column;
-  @media screen and (max-width: 768px) {
-    width: 100%;
-  }
-  @media screen and (max-width: 520px) {
-    width: 100%;
-  }
-  div {
-    width: 20.8rem;
-    height: 20.8rem;
-    object-fit: cover;
-    justify-content: center;
-    border-radius: 0.6rem;
-    @media screen and (max-width: 768px) {
-      width: 14rem;
-      height: 14rem;
-      margin-bottom: 1rem;
-    }
-  }
-
-  img {
-    object-fit: cover;
-    object-position: center;
-    width: 100%;
-    height: 100%;
-    border-radius: 0.6rem;
-    border-style: none;
-    @media screen and (max-width: 768px) {
-      width: 14rem;
-      height: 14rem;
-      border-radius: 0.6rem;
-    }
-  }
-
-  h3 {
-    font-size: var(--fontSize-body);
-    color: var(--11-gray);
-    margin-top: 1rem;
-    overflow: hidden;
-    display: -webkit-box;
-    -webkit-line-clamp: 1;
-    -webkit-box-orient: vertical;
-    white-space: normal;
-    overflow: hidden;
-    height: 2.5rem;
-    white-space: wrap;
-    text-overflow: ellipsis;
-    line-height: 1.5;
-    @media screen and (max-width: 768px) {
-      width: 14rem;
-      height: 2rem;
-      margin-top: 0.6rem;
-      color: var(--11-gray, #f8f8f8);
-      font-weight: var(--fontWeight-medium);
-      font-size: var(--fontSize-H5);
-    }
-  }
-
-  p {
-    font-size: var(--fontSize-body);
-    font-weight: var(--fontWeight-bold);
-    color: var(--11-gray);
-    margin-top: 1rem;
-    text-align: left;
-    @media screen and (max-width: 768px) {
-      height: 2.3rem;
-      font-weight: var(--fontWeight-bold);
-      font-size: var(--fontSize-H5);
-    }
-  }
-`;
-
-interface QualityProps {
-  $quality: string;
-}
-
-const ProductsCardQuality = styled.h1<QualityProps>`
-  width: 9rem;
-  padding: 0 0.8rem;
-  text-align: center;
-  line-height: 1.7;
-  border-radius: 0.3rem;
-  margin-top: 1rem;
-  background-color: #fcfcfc;
-
-  color: var(--2-gray);
-  margin-bottom: 0.6rem;
-  font-size: var(--fontSize-H6);
-  // 배경색 조건부 렌더링
-  ${(props) => {
-    if (props.children === '새상품(미사용)') {
-      return css`
-        background-color: var(--opc-100);
-        color: var(--2-gray);
-      `;
-    }
-    if (props.children === '고장/파손 상품') {
-      return css`
-        background-color: var(--4-gray);
-        color: var(--11-gray);
-      `;
-    }
-  }}
-  @media screen and (max-width: 768px) {
-    margin-top: 1rem;
-    width: 8rem;
-    height: 2rem;
-    font-size: 1rem;
-    font-weight: 500;
-    line-height: 191.2%;
-    text-align: center;
-  }
-`;
-
-const CommunityResultsContainer = styled.div<{
-  showClickedList: boolean;
-}>`
-  width: 77.5%;
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  margin: 0 auto;
-  margin-top: 5rem;
-  @media screen and (max-width: 768px) {
-    width: 93%;
-    margin-top: ${({ showClickedList }) => (showClickedList ? 0 : '2rem')};
-    padding: 0 1rem;
-  }
-  @media screen and (max-width: 530px) {
-    width: 100%;
-    padding: 0 1rem;
-    margin-top: ${({ showClickedList }) => (showClickedList ? 0 : '2rem')};
-  }
-  @media screen and (max-width: 349px) {
-    width: 100%;
-    margin-top: ${({ showClickedList }) => (showClickedList ? 0 : '2rem')};
-    padding: 0 1rem;
-  }
-`;
-const CommunityTitle = styled.div<{
-  showClickedList: boolean;
-}>`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-  margin: 0 auto;
-  margin-bottom: 2.2rem;
-  @media screen and (max-width: 768px) {
-    display: none;
-    margin-bottom: 1rem;
-  }
-  @media screen and (max-width: 530px) {
-    display: none;
-    margin-bottom: 1rem;
-  }
-  @media screen and (max-width: 349px) {
-    display: none;
-    margin-bottom: 1rem;
-  }
-`;
-const LinktoCommunityPosts = styled.div`
-  text-decoration: none;
-  cursor: pointer;
-  font-weight: bold;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 8.3rem;
-  height: 3.2rem;
-  text-decoration: none;
-  cursor: pointer;
-  font-weight: var(--fontWeight-bold);
-  gap: 0.8rem;
-  background: var(--opc-20);
-  border-radius: 4.5rem;
-  font-size: var(--fontSize-H6);
-  font-weight: var(--fontWeight-medium);
-  color: var(--11-gray);
-  @media screen and (max-width: 768px) {
-    display: flex;
-    background: none;
-    width: 6rem;
-    font-size: 1.1rem;
-    gap: 0.3rem;
-  }
-  &:hover {
-    background-color: var(--opc-100);
-    color: var(--bgColor);
-  }
-  svg {
-    width: 1rem;
-    height: 0.9rem;
-    @media screen and (max-width: 768px) {
-      width: 9px;
-      height: 8px;
-      color: var(--opc-100);
-    }
-  }
-`;
-const CommunityPostsList = styled.ul`
-  width: 100%;
-  margin-top: 2.2rem;
-  background-color: transparent;
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 1.5rem;
-  @media screen and (max-width: 1200px) {
-    grid-template-columns: repeat(2, 1fr);
-    margin: 0 auto;
-    /* margin-top */
-  }
-
-  @media screen and (max-width: 768px) {
-    grid-template-columns: 1fr;
-    padding: 0 1rem;
-  }
-
-  @media screen and (max-width: 520px) {
-    grid-template-columns: 1fr;
-    padding: 0 1rem;
-  }
-`;
-const ToCommunityPage = styled(Link)`
-  text-decoration: none;
-  cursor: pointer;
-  color: var(--11-gray);
-  @media screen and (max-width: 768px) {
-    width: 100%;
-  }
-`;
-const PostList = styled.li`
-  width: 54.6rem;
-  height: 19.5rem;
-  display: inline-block;
-  position: relative;
-  align-items: center;
-  border-radius: 1rem;
-  background-color: var(--3-gray);
-  padding: 2rem;
-  margin-bottom: 2rem;
-
-  @media screen and (max-width: 768px) {
-    width: 100%;
-    padding: 1rem;
-  }
-  @media screen and (max-width: 520px) {
-    width: 100%;
-    padding: 1rem;
-  }
-  .nopicture {
-    width: 6.6rem;
-    height: 6.6rem;
-    object-fit: cover;
-    @media screen and (max-width: 768px) {
-      width: 4rem;
-      height: 4rem;
-    }
-  }
-  .commupic {
-    display: flex;
-    gap: 1.2rem;
-    @media screen and (max-width: 768px) {
-    }
-  }
-  .commucontent {
-    /* margin-left: 1.5rem; */
-    /* margin-bottom: 3rem; */
-    /* gap: 10px; */
-  }
-  .community-pic {
-    width: 6.6rem;
-    height: 6.6rem;
-    object-fit: cover;
-    @media screen and (max-width: 768px) {
-      width: 4rem;
-      height: 4rem;
-    }
-  }
-  h3 {
-    color: var(--11-gray);
-    font-size: var(--fontSize-H4);
-    margin-bottom: 1.6rem;
-    font-weight: var(--fontWeight-bold);
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    max-width: 48rem;
-    @media screen and (max-width: 768px) {
-      margin-top: 1rem;
-      font-size: var(--fontSize-H5);
-      font-weight: var(--fontWeight-bold);
-      width: 20rem;
-    }
-  }
-
-  p {
-    display: -webkit-box;
-    -webkit-line-clamp: 3;
-    -webkit-box-orient: vertical;
-    white-space: normal;
-    overflow: hidden;
-    font-size: var(--fontSize-H4);
-    font-weight: var(--fontWeight-medium);
-    color: var(--8-gray);
-    height: 6.6rem;
-    line-height: 1.2;
-
-    @media screen and (max-width: 768px) {
-      font-size: var(--fontSize-H6);
-      -webkit-line-clamp: 2;
-      line-height: 1.92;
-      height: 4rem;
-    }
-  }
-
-  .thumbs {
-    position: absolute;
-    bottom: 1.5rem;
-    left: 3.5rem;
-    width: 2rem;
-    height: 2rem;
-    @media screen and (max-width: 768px) {
-      width: 1.3rem;
-      height: 1.2rem;
-      left: 3rem;
-      bottom: 1.5rem;
-    }
-  }
-  .likescount {
-    position: absolute;
-    text-decoration: none;
-    bottom: 1.5rem;
-    left: 7rem;
-    color: var(--6, #717171);
-    @media screen and (max-width: 768px) {
-      font-size: 1.1rem;
-      left: 5.5rem;
-    }
-  }
-  .commentss {
-    position: absolute;
-    bottom: 1.5rem;
-    left: 12rem;
-    width: 2rem;
-    height: 2rem;
-    @media screen and (max-width: 768px) {
-      width: 1.3rem;
-      height: 1.2rem;
-      left: 8.5rem;
-      bottom: 1.5rem;
-    }
-  }
-
-  span {
-    position: absolute;
-    text-decoration: none;
-    bottom: 1.5rem;
-    left: 15rem;
-    color: var(--6, #717171);
-    @media screen and (max-width: 768px) {
-      font-size: 1.1rem;
-      left: 11rem;
-    }
-  }
-  h4 {
-    position: absolute;
-    bottom: 1.5rem;
-    right: 1.5rem;
-    color: var(--6, #717171);
-    font-size: var(--fontSize-H6);
-    @media screen and (max-width: 768px) {
-      font-size: 1rem;
-      right: 1.5rem;
-    }
-  }
-`;
